- Refactor visual query builder - Implement metadata select filter
